logic refactoring
