+++ src/App.scss
@@ -44,3 +44,7 @@
   flex-direction: column;
   width: 100vw;
 }
+
+p {
+  margin-bottom: 0 !important;
+}
