+++ src/components/Menu/index.scss
@@ -23,11 +23,18 @@
     padding-inline-start: 0;
   }
   &__list__item {
+    cursor: pointer;
     display: flex;
     text-transform: uppercase;
     color: $white;
+    height: 80px;
+    align-items: center;
+    &__active {
+      border-left: 4px solid $white;
+    }
   }
 }
+
 .list_item__logo {
   align-self: center;
   margin-left: 30px;
