+++ src/components/User/index.scss
@@ -1,4 +1,4 @@
-@import "../../variables.scss";
+@import "../variables.scss";
 
 .customer_email,
 .customer_name,
