+++ src/components/UserPopup/index.scss
@@ -1,4 +1,4 @@
-@import "../../variables.scss";
+@import "../variables.scss";
 
 .user_popup__container {
   display: flex;
