+++ src/components/variables.scss
